requirements.txt
streamlit
baserun
vertexai
pandas
mysql-connector-python
sendgrid
argparse
os
time
json